Refactor and ComponentScan
