+++ src/main/java/com/viktor/controller/WorldController.java
@@ -1,6 +1,8 @@
-package com.viktor.main;
+package com.viktor.controller;
 
 import java.util.concurrent.atomic.AtomicLong;
+
+import com.viktor.model.World;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
