--- src/main/java/com/viktor/main/NunegalEntrevistaApplication.java
@@ -1,13 +1,0 @@
-package com.viktor.main;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class NunegalEntrevistaApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(NunegalEntrevistaApplication.class, args);
-	}
-
-}
