+++ src/main/java/com/viktor/main/TestApplication.java
@@ -1,0 +1,15 @@
+package com.viktor.main;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.ComponentScan;
+
+@SpringBootApplication
+@ComponentScan({"com.viktor.main","com.viktor.controller","com.viktor.model"})
+public class TestApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(TestApplication.class, args);
+	}
+
+}
